split the stream into a receive and a send stream
