+++ crypto_stream_test.go
@@ -14,8 +14,8 @@ var _ = Describe("Stream", func() {
 
 	It("sets the read offset", func() {
 		str.SetReadOffset(0x42)
-		Expect(str.readOffset).To(Equal(protocol.ByteCount(0x42)))
-		Expect(str.frameQueue.readPosition).To(Equal(protocol.ByteCount(0x42)))
+		Expect(str.receiveStream.readOffset).To(Equal(protocol.ByteCount(0x42)))
+		Expect(str.receiveStream.frameQueue.readPosition).To(Equal(protocol.ByteCount(0x42)))
 	})
 
 	It("says if it has data for writing", func() {
